requirements.txt
discord.py==1.7.3
async-timeout==3.0.1
dotnet==1.4.2
giphy_client=1.0.0
dnspython==1.16.0
PyNaCl==1.3.0
git+https://github.com/Rapptz/discord.py@rewrite
